file structure and in progress cd_cmd
