+++ src/executor/builtins/env_cmd.c
@@ -1,0 +1,12 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   env_cmd.c                                          :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: aabelkis <aabelkis@student.42.fr>          +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2025/09/10 15:02:43 by aabelkis          #+#    #+#             */
+/*   Updated: 2025/09/10 15:02:44 by aabelkis         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
