+++ src/executor/builtins/exir_cmd.c
@@ -1,0 +1,12 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   exir_cmd.c                                         :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: aabelkis <aabelkis@student.42.fr>          +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2025/09/10 15:02:49 by aabelkis          #+#    #+#             */
+/*   Updated: 2025/09/10 15:02:50 by aabelkis         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
