+++ src/executor/builtins/export_cmd.c
@@ -1,0 +1,12 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   export_cmd.c                                       :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: aabelkis <aabelkis@student.42.fr>          +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2025/09/10 15:02:54 by aabelkis          #+#    #+#             */
+/*   Updated: 2025/09/10 15:02:55 by aabelkis         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
