+++ src/executor/builtins/pwd_cmd.c
@@ -1,0 +1,12 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   pwd_cmd.c                                          :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: aabelkis <aabelkis@student.42.fr>          +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2025/09/10 15:02:59 by aabelkis          #+#    #+#             */
+/*   Updated: 2025/09/10 15:03:00 by aabelkis         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
