+++ src/executor/builtins/unset_cmd.c
@@ -1,0 +1,13 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   unset_cmd.c                                        :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: aabelkis <aabelkis@student.42.fr>          +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2025/09/10 15:03:05 by aabelkis          #+#    #+#             */
+/*   Updated: 2025/09/10 15:03:06 by aabelkis         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+env
